requirements.txt
farm-haystack transformers ollama
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
"farm-haystack[inference]"

